Added API_SECRET for deletion
Added a constant API_SECRET for deletion.

+++ api/views.py
@@ -1,10 +1,9 @@
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
+from django.conf import settings
 from rest_framework.parsers import JSONParser
 from rest_framework.decorators import api_view, renderer_classes
-
 from rest_framework import response
-
 from models import *
 from serializers import *
 
@@ -45,6 +44,14 @@ def user_detail(request, pk, format=None):
         return response.Response(serializer.data)
 
     elif request.method == 'DELETE':
+        if 'HTTP_API_SECRET' not in request.META:
+            data = {}
+            data["details"] = "api-secret were not provided."
+            return response.Response(status=401, data=data)
+        if reques.META['HTTP_API_SECRET'] != settings.API_SECRET_KEY:
+            data = {}
+            data["details"] = "Invalid api-secret."
+            return response.Response(status=401, data=data)
 
         user.delete()
         return response.Response(status=204)
